GH-5199: Organized couple of traits into separate module

+++ smart_contracts/contracts/vm2/vm2-harness/src/contracts/no_fallback.rs
@@ -5,8 +5,8 @@ use casper_sdk::casper;
 #[derive(Default)]
 #[casper(contract_state)]
 pub struct NoFallback {
-    initial_balance: u128,
-    received_balance: u128,
+    initial_balance: u64,
+    received_balance: u64,
 }
 
 #[casper]
